+++ sockeye/translate.py
@@ -76,9 +76,7 @@ def run_translate(args: argparse.Namespace):
                                                        dtype=args.dtype,
                                                        clamp_to_dtype=args.clamp_to_dtype,
                                                        inference_only=True,
-                                                       knn_index=args.knn_index,
-                                                       knn_cache_size=args.knn_cache_size,
-                                                       knn_cache_alpha=args.knn_cache_alpha)
+                                                       knn_index=args.knn_index)
 
     restrict_lexicon = None  # type: Optional[Union[RestrictLexicon, Dict[str, RestrictLexicon]]]
     if args.restrict_lexicon is not None:
@@ -114,8 +112,6 @@ def run_translate(args: argparse.Namespace):
     else:
         raise ValueError("Unknown brevity penalty type %s" % args.brevity_penalty_type)
 
-    knn_lambda = args.knn_lambda
-
     for model in models:
         model.eval()
 
@@ -140,7 +136,7 @@ def run_translate(args: argparse.Namespace):
                                       sample=args.sample,
                                       output_scores=output_handler.reports_score(),
                                       constant_length_ratio=constant_length_ratio,
-                                      knn_lambda=knn_lambda,
+                                      knn_lambda=args.knn_lambda,
                                       max_output_length_num_stds=args.max_output_length_num_stds,
                                       max_input_length=args.max_input_length,
                                       max_output_length=args.max_output_length,